==> src/01_fishbase_data_wrangling.Rmd <==
---
title: "fish_base"
author: "Leonardo Feitosa"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(rfishbase)
library(here)
library(janitor)
library(vroom)
```



```{r}
# get shark and ray scientific names and their biological parameters from FishBase
shark_spp <- fb_tbl("species") %>%
  clean_names() %>%
  filter(!str_detect(genus, "Potamotrygon|Paratrygon|Heliotrygon|Plesiotrygon|Glyphis")) %>% # remove freshwater elasmos
  select(genus, species, f_bname, demers_pelag) %>%
  unite(scientific_name, c("genus", "species"), sep = " ", remove = FALSE) %>%
  mutate(common_name = str_to_lower(f_bname))
```

```{r}
# code to create the cleaned version of the raw dataset
# call PRM dataset
prm_elasmo <- read_csv(here("data", "prm_dataset_updated.csv"))

# prm_elasmo <- prm_elasmo %>%
#   rename(scientific_name = species)

prm_elasmo_join <- prm_elasmo %>%
  left_join(shark_spp, by = "scientific_name") %>%
  select(-habitat_associated) %>%
  rename(habitat_associated = demers_pelag) %>%
  relocate(habitat_associated, .after = median_temp) %>%
  mutate(habitat_associated = case_when(
    habitat_associated == "pelagic-oceanic" ~ "pelagic",
    scientific_name == "Prionace glauca" ~ "pelagic",
    str_detect(scientific_name, "canis") ~ "demersal",
    str_detect(scientific_name, "Centrophorus") ~ "bathydemersal",
    str_detect(scientific_name, "perlo") ~ "bathydemersal",
    str_detect(scientific_name, "Deania") ~ "bathydemersal",
    str_detect(scientific_name, "Himantura") ~ "demersal",
    str_detect(scientific_name, "Negaprion") ~ "reef-associated",
    str_detect(scientific_name, "limbatus") ~ "reef-associated",
    str_detect(scientific_name, "Leucoraja") ~ "demersal",
    str_detect(scientific_name, "Rhizo") ~ "demersal",
    str_detect(scientific_name, "antarticus") ~ "demersal",
    str_detect(scientific_name, "Pseudo") ~ "pelagic",
    str_detect(scientific_name, "Centroscymnus owstoni") ~ "bathydemersal",
    str_detect(scientific_name, "Trygonoptera testacea") ~ "demersal",
    TRUE ~ habitat_associated
  )) %>%
  mutate(scientific_name = case_when(
    str_detect(scientific_name, "antarticus") ~ "Mustelus antarcticus",
    str_detect(scientific_name, "Rhizoprionodon terranovae") ~ "Rhizoprionodon terraenovae",
    str_detect(scientific_name, "Squalus sp.") ~ "Squalus sp",
    str_detect(scientific_name, "Orectolobus spp.") ~ "Orectolobus sp",
    TRUE ~ scientific_name
  )) %>%
  mutate(family = ifelse(scientific_name == "Rhynchobatus australiae", "Rhinidae", family))

write_csv(prm_elasmo_join, file = here("data", "prm_dataset_updated.csv"), col_names = TRUE, na = "")
```

```{r}
# create a data subset with only sharks and variables of interest
prm_elasmo_subset <- read_csv(here("data", "prm_dataset_updated.csv")) %>% 
  mutate(group = case_when(
    str_detect(family, "Mobulidae|Dasyatidae|Gymnuridae|Myliobatidae|Torpedinidae|Rhinobatidae|Rhinidae|Aetobatidae|Rajidae|Pristidae|Plesiobatidae|Potamo|Urolophidae|Zanobatidae|Narcinidae|Platyrhinidae|Trygon|Hypnidae|Narkidae") ~ "Batoids",
    str_detect(
      scientific_name,
      "Himantura|Dasyatis|trygon|Bathytoshia|rays|raja|Rhinoptera|Sympterygia|Pastinachus|Urobatis|Glaucostegus|Hypanus|Rhynchobatus"
    ) ~ "Batoids",
    TRUE ~ "Sharks"
  )) %>%
  relocate(group, .after = family) %>%
  mutate(
    scientific_name = ifelse(scientific_name == "Mustelus canis insularis", "Mustelus canis", scientific_name),
    scientific_name = ifelse(scientific_name == "Centroscymnus owstoni", "Centroscymnus owstonii", scientific_name)
  ) %>%
  filter(!scientific_name %in% c("sharks", "rays")) %>%
  select(scientific_name, family, group, gear_class, habitat_associated, estimate_type, estimate, sample_size, method, ventilation_method, max_size_cm, median_depth, reproductive_mode, reference)

write_csv(prm_elasmo_subset, file = here("data", "prm_elasmo_subset.csv"), col_names = TRUE, na = "")
```


```{r}
# code to create the final dataset of population growth rates
# call the intrinsic population growth rate from Bradley et al (XXXX)
r_data <- read_csv(here("data", "r_estimate_fish.csv")) %>%
  rename(scientific_name = SciName) %>%
  pivot_longer(
    cols = r_fin,
    names_to = "pop_growth_rate_measure",
    values_to = "value"
  )

# call in the intrinsic growth rate data and wrangle it to convert lambda to r
int_growth <- read_csv(here("data", "intrinsic_pop_growth_rates.csv"))

# write_csv(int_growth, file = here("data", "intrinsic_pop_growth_rates.csv"), col_names = TRUE, na = "")

# build final dataset with intrinsic population growth rates for each species 
r_data_full <- bind_rows(r_data, int_growth) %>%
  drop_na(pop_growth_rate_measure) %>%
  select(scientific_name, pop_growth_rate_measure, value, areas, reference, observations) %>%
  mutate(r_value = case_when(
    pop_growth_rate_measure == "lambda" ~ log(value),
    TRUE ~ value
  )) %>%
  group_by(scientific_name) %>%
  mutate(mean_r = mean(r_value)) %>% # calculates mean r values per species because some species have different r values depending on method/sampling area
  relocate(areas, .after = scientific_name) %>%
  relocate(reference, .after = mean_r) %>%
  relocate(observations, .after = reference) %>%
  distinct()

# write_csv(r_data_full, file = here("data", "intrinsic_pop_growth_rates_full.csv"), col_names = TRUE, na = "")
```

```{r}
# Get list of species with longline or gillnet listed as a threat in their IUCN assessments
# iucn data
iucn_taxonomy <- read_csv(here("data", "iucn_data", "taxonomy.csv")) %>%
  clean_names() %>%
  select(scientific_name, family_name) %>%
  rename(family = family_name) %>%
  mutate(family = str_to_sentence(family))

iucn_assessment_long <- read_csv(here("data", "iucn_data", "assessments.csv")) %>%
  clean_names() %>%
  filter(str_detect(systems, "Marine") & str_detect(threats, "longline") | str_detect(scientific_name, "Squatina|Isogomphodon|Carcharhinus|Eusphyra|Orectolobus|Pristiophorus|Mustelus")) %>%
  select(scientific_name, threats, systems, redlist_category)

iucn_assessment_gill <- read_csv(here("data", "iucn_data", "assessments.csv")) %>%
  clean_names() %>%
  filter(str_detect(systems, "Marine") & str_detect(threats, "gillnet")) %>%
  select(scientific_name, threats, systems, redlist_category)

iucn_data_long <- iucn_assessment_long %>%
  left_join(iucn_taxonomy, by = "scientific_name") %>%
  filter(!str_detect(family, "Mobulidae|Dasyatidae|Gymnuridae|Myliobatidae|Torpedinidae|Rhinobatidae|Rhinidae|Aetobatidae|Rajidae|Pristidae|Plesiobatidae|Potamo|Urolophidae|Zanobatidae|Narcinidae|Platyrhinidae|Trygon|Hypnidae|Narkidae|Gurgesiellidae|Rhinopteridae|Urotrygonidae|Arhynchobatidae|Glaucostegidae|Anacanthobatidae|himaeridae|Callorhinchidae|Dichichthyidae"))

# write_csv(iucn_data_long, file = here("data", "predict_longline.csv"))

iucn_data_gill <- iucn_assessment_gill %>%
  left_join(iucn_taxonomy, by = "scientific_name") %>%
  filter(!str_detect(family, "Mobulidae|Dasyatidae|Gymnuridae|Myliobatidae|Torpedinidae|Rhinobatidae|Rhinidae|Aetobatidae|Rajidae|Pristidae|Plesiobatidae|Potamo|Urolophidae|Zanobatidae|Narcinidae|Platyrhinidae|Trygon|Hypnidae|Narkidae|Gurgesiellidae|Rhinopteridae|Urotrygonidae|Arhynchobatidae|Glaucostegidae|Anacanthobatidae|himaeridae|Callorhinchidae|Dichichthyidae")) %>%
  filter(!str_detect(scientific_name, "Glyphis"))

# write_csv(iucn_data_gill, file = here("data", "predict_gillnet.csv"))


```

### Check hypoxia data

```{r}
# wrangle out literature review data with Gilman et al (2022)'s and Penn & Deustch (2024)'s data
prm_elasmo <- read_csv(here("data", "prm_elasmo_subset.csv"))


new_dat <- read_csv(here::here("data", "predict_data.csv")) %>%
  filter(!is.na(median_depth)) %>%
  filter(!is.na(ventilation_method)) %>%
  filter(!is.na(reproductive_mode)) %>%
  filter(!is.na(habitat_associated)) %>%
  filter(!is.na(max_size_cm)) %>%
  filter(!str_detect(scientific_name, "/")) %>%
  mutate(ventilation_method = as.factor(ventilation_method)) %>%
  mutate(habitat_associated = as.factor(habitat_associated)) %>%
  mutate(reproductive_mode = as.factor(reproductive_mode)) %>%
  filter(reproductive_mode != "histotrophy") %>%
  mutate(
    scientific_name = ifelse(scientific_name == "Mustelus canis insularis", "Mustelus canis", scientific_name),
    scientific_name = ifelse(scientific_name == "Centroscymnus owstoni", "Centroscymnus owstonii", scientific_name)
  )

# read in the longline threat data and join with longline predict data
iucn_data_long <- read_csv(here("data", "predict_longline.csv"))

# join the datasets
new_dat_longline <- iucn_data_long %>%
  left_join(new_dat, by = c("scientific_name", "family")) %>%
  distinct(
    scientific_name, family, redlist_category, habitat_associated,
    ventilation_method, max_size_cm, median_depth, reproductive_mode
  )

# read in the longline threat data and join with gillnet predict data
iucn_data_gill <- read_csv(here("data", "predict_gillnet.csv"))

# join the datasets
new_dat_gill <- iucn_data_gill %>%
  left_join(new_dat, by = c("scientific_name", "family")) %>%
  distinct(
    scientific_name, family, redlist_category, habitat_associated,
    ventilation_method, max_size_cm, median_depth, reproductive_mode
  )

# create subset of the new_dat to join for the median depth, ventilation and reproductive modes
new_dat_long_subset <- new_dat_longline %>%
  select(scientific_name, family, max_size_cm, median_depth, reproductive_mode, ventilation_method)

tol <- read_csv(here("data", "tolerance.csv")) %>%
  rename(scientific_name = species) %>%
  mutate(scientific_name = ifelse(scientific_name == "Stegostoma fasciatum", "Stegostoma tigrinum", scientific_name)) %>%
  filter(!is.na(ac)) %>%
  select(-eo)
# eo is thermal sensitivity
# ac is active hypoxia vulnerability

# read in the Gilman et al (2022) data
gilman_raw <- read_csv(here("data", "Gilman_et_2022_elasmobranch-longline-haulbak-conditon_database_PUBLIC.xlsx - AVMdatabase.csv")) %>%
  clean_names() %>%
  select(-c(pub_uid, subset, region, studycat, soak_max_hours, branchline_length_m, hooks_between_floats)) %>%
  # filter(!genus %in% c("Pteroplatytrygon", "Mobula")) %>%
  mutate(source = "Gilman et al (2022)") %>%
  rename(common_name = commonname) %>%
  mutate(common_name = str_replace(common_name, "shark", "")) %>%
  mutate(common_name = str_trim(common_name, side = "right")) %>%
  mutate(common_name = ifelse(common_name == "shorfin mako", "shortfin mako", common_name)) %>%
  mutate(genus = ifelse(genus == "Rhinobatos", "Glaucostegus", genus)) %>%
  mutate(family = ifelse(family == "Rhinobatidae", "Glaucostegidae", family)) %>%
  mutate(group = case_when(
    str_detect(family, "Mobulidae|Dasyatidae|Rhinobatidae|Rajidae|Glaucostegidae|Myliobatidae") ~ "Batoids",
    TRUE ~ "Sharks"
  )) %>%
  group_by(record_uid) %>%
  mutate(sample_size = sum(alive, dead)) %>%
  ungroup() %>%
  mutate(estimate = dead / (alive + dead)) %>%
  select(-alive, -dead)

# wrangle the shark_spp df
shark_spp_subset <- shark_spp %>%
  select(scientific_name, common_name, genus) %>%
  mutate(common_name = str_replace(common_name, "shark", "")) %>%
  mutate(common_name = str_trim(common_name, side = "right")) %>%
  mutate(scientific_name = ifelse(scientific_name == "Carcharhinus perezii", "Carcharhinus perezi", scientific_name)) %>%
  mutate(common_name = ifelse(common_name == "giant manta", "giant manta ray", common_name))

gilman_join <- gilman_raw %>%
  left_join(shark_spp_subset, by = c("genus", "common_name")) %>%
  mutate(scientific_name = case_when(
    genus == "Pseudocarcharias" ~ "Pseudocarcharias kamoharai",
    genus == "Scymnodon" ~ "Scymnodon ichiharai",
    common_name == "grey reef" ~ "Carcharhinus amblyrhynchos",
    genus == "Heptranchias" ~ "Heptranchias perlo",
    common_name == "sicklefin lemon" ~ "Negaprion acutidens",
    common_name == "lemon" ~ "Negaprion brevirostris",
    TRUE ~ scientific_name
  )) %>%
  mutate(gear_class = "longline") %>%
  left_join(shark_spp, by = "scientific_name") %>%
  rename(habitat_associated = demers_pelag)
```

```{r}
dat <- full_join(prm_elasmo, new_dat_longline) %>%
  mutate(species = scientific_name) %>%
  separate_wider_delim(species, names = c("genus", "species"), delim = " ") %>%
  mutate(source = "this study")

avm_gilman_hypoxia <- gilman_join %>%
  #  select(- c(max_size_cm, l_type_max_m, depth_range_shallow, depth_range_deep)) %>%
  left_join(new_dat_long_subset, by = c("scientific_name", "family")) %>%
  bind_rows(dat) %>%
  mutate(estimate_type = ifelse(source == "Gilman et al (2022)", "at-vessel mortality", estimate_type)) %>%
  select(-c(record_uid, common_name.x, common_name.y, genus.y, -f_bname)) %>%
  left_join(tol) %>%
  select(-bait, -hooks_per_set, -f_bname, -species, -method) %>%
  mutate(genus = case_when(
    !is.na(genus) ~ genus,
    is.na(genus) ~ genus.x
  )) %>%
  select(-genus.x) %>%
  mutate(max_size_cm = case_when(
    str_detect(scientific_name, "Ptero") ~ 163.5,
    str_detect(scientific_name, "birostris") ~ 700,
    str_detect(scientific_name, "tarapacana") ~ 370,
    str_detect(scientific_name, "grabatus") ~ 150,
    str_detect(scientific_name, "cemiculus") ~ 245,
    str_detect(scientific_name, "clavata") ~ 130,
    str_detect(scientific_name, "bovinus") ~ 222,
    str_detect(scientific_name, "japanica") ~ 314,
    str_detect(scientific_name, "mobular") ~ 520,
    # scientific_name == "Mustelus canis" ~ 155,
    # scientific_name == "Mustelus antarcticus" ~ 185.2,
    TRUE ~ max_size_cm
  )) %>%
  mutate(median_depth = case_when(
    str_detect(scientific_name, "Ptero") ~ 190.5,
    str_detect(scientific_name, "birostris") ~ 500,
    str_detect(scientific_name, "tarapacana") ~ 948,
    str_detect(scientific_name, "grabatus") ~ 50,
    str_detect(scientific_name, "cemiculus") ~ 54.5,
    str_detect(scientific_name, "clavata") ~ 513,
    str_detect(scientific_name, "bovinus") ~ 50,
    str_detect(scientific_name, "japanica") ~ 324.5,
    str_detect(scientific_name, "mobular") ~ 350,
    # scientific_name == "Mustelus canis" ~ 404,
    # scientific_name == "Mustelus antarcticus" ~ 175.5,
    # scientific_name == "Galeocerdo cuvier" ~ 556,
    # scientific_name == "Centroscymnus owstonii" ~ 804.5,
    TRUE ~ median_depth
  )) %>%
  mutate(ventilation_method = case_when(
    str_detect(scientific_name, "Ptero|grabatus|cemiculus|clavata|bovinus") ~ "stationary",
    str_detect(scientific_name, "tarapacana|birostris|mobular|japanica") ~ "ram",
    TRUE ~ ventilation_method
  )) %>%
  distinct()

avm_gilman_hypoxia_check <- avm_gilman_hypoxia %>%
  filter(is.na(max_size_cm))
```

```{r}
# populate NAs in ac with genus or family averages
dat_new_mean <- avm_gilman_hypoxia %>%
  group_by(genus) %>%
  mutate(ac_new_gen = ifelse(is.na(ac), mean(ac, na.rm = TRUE), ac)) %>%
  group_by(family) %>%
  mutate(ac_new_fam = ifelse(is.na(ac), mean(ac, na.rm = TRUE), ac)) %>%
  ungroup() %>%
  mutate(
    ac = ifelse(is.na(ac), ac_new_gen, ac),
    ac = ifelse(is.nan(ac_new_gen), ac_new_fam, ac)
  ) %>%
  mutate(ac = ifelse(scientific_name == "Glaucostegus cemiculus", 5.7727, ac)) %>% # using the value for G. typus from the 'tol' df
  filter(!is.na(ac)) %>%
  select(-ac_new_gen, -ac_new_fam)

dat_predict <- dat_new_mean %>%
  filter(!is.na(estimate)) %>% 
  mutate(source = ifelse(source == "this study", reference, source)) %>% 
  select(-reference)

write_csv(dat_predict, here::here("data", "prm_hypoxia.csv"))

dat_new_mean_check <- dat_new_mean %>%
  filter(scientific_name %in% new_dat$scientific_name) %>%
  filter(is.na(estimate))


write_csv(dat_new_mean, here::here("data", "predict_data_hyp.csv"), col_names = TRUE, na = "")
```

```{r}

```

```{r}
# testing the reproductive mode data with Mull et al (2024)'s
mull_df <- read_csv(here("data", "mull_etal_reproduction.csv")) %>%
  clean_names() %>%
  rename(scientific_name = species) %>%
  mutate(scientific_name = str_replace(scientific_name, "_", " ")) %>%
  select(-reproductive_mode) %>%
  mutate(reproductive_mode = str_to_lower(vertebrate_reproductive_category)) %>%
  drop_na() %>%
  select(scientific_name, reproductive_mode)

prm_data <- read_csv(here("data", "prm_hypoxia.csv")) %>%
  mutate(reproductive_mode = case_when(
    reproductive_mode == "placental viviparity" ~ "matrotrophic viviparity",
    reproductive_mode == "yolk-sac viviparity" ~ "lecithotrophic viviparity",
    reproductive_mode == "oophagy" ~ "matrotrophic viviparity",
    reproductive_mode == "oviparous" ~ "lecithotrophic oviparity",
    reproductive_mode == "histotrophy" ~ "matrotrophic viviparity"
  ))
```

```{r}
# join ours and Mull et al (2024)'s data
mull_join <- prm_data %>%
  left_join(mull_df, by = c("scientific_name", "reproductive_mode")) %>%
  mutate(habitat_associated = ifelse(scientific_name == "Carcharhinus perezi", "reef-associated", habitat_associated)) %>%
  mutate(reproductive_mode = case_when(
    str_detect(scientific_name, "Ptero|Mobula|bovinus|Taeniurops") ~ "matrotrophic viviparity",
    str_detect(scientific_name, "clavata") ~ "lecithotrophic oviparity",
    str_detect(scientific_name, "cemiculus") ~ "lecithotrophic viviparity",
    TRUE ~ reproductive_mode
  )) %>% 
  mutate(reproductive_mode = case_when(
    scientific_name == "Cetorhinus maximus" ~ "matrotrophic viviparity",
    scientific_name == "Rhincodon typus" ~ "lecithotrophic viviparity",
    TRUE ~ reproductive_mode
  )) 

# update dataset for model fitting
write_csv(mull_join, here::here("data", "prm_hypoxia.csv"), na = "")

mull_join <- mull_join %>%  
  mutate(habitat = case_when(
    str_detect(habitat_associated, "pelagic") ~ "pelagic",
    str_detect(habitat_associated, "reef") ~ "pelagic",
    str_detect(habitat_associated, "demersal") ~ "demersal"
  )) %>% 
  select(-habitat_associated)

## create df for Dryad submission
write_csv(mull_join, here::here("data", "elasmobranch_avm_prm_lit_review_final.csv"), na = "")
```

```{r}
predict_data <- read_csv(here("data", "predict_data_hyp.csv")) %>%
  mutate(reproductive_mode = case_when(
    reproductive_mode == "placental viviparity" ~ "matrotrophic viviparity",
    reproductive_mode == "yolk-sac viviparity" ~ "lecithotrophic viviparity",
    reproductive_mode == "oophagy" ~ "matrotrophic viviparity",
    reproductive_mode == "oviparous" ~ "lecithotrophic oviparity",
    reproductive_mode == "histotrophy" ~ "matrotrophic viviparity"
  ))

# merge predict data with Mull et al (2024)'s data
mull_join_predict <- predict_data %>%
  left_join(mull_df, by = c("scientific_name", "family", "reproductive_mode"))

# update predict data
write_csv(mull_join_predict, here::here("data", "predict_data_hyp.csv"), col_names = TRUE, na = "")
```

```{r}
# join the predict longline and gillnet data to check for missing species
predict_data_long <- read_csv(here::here("data", "predict_data_hyp.csv"))

prm_hypoxia <- read_csv(here("data", "prm_hypoxia.csv")) %>%
  select(scientific_name, family, ventilation_method, reproductive_mode, median_depth, max_size_cm, habitat_associated, ac) %>%
  distinct()


predict_gillnet <- read_csv(here::here("data", "predict_gillnet.csv"))

predict_gillnet_join <- predict_gillnet %>%
  full_join(predict_data_long, by = c("scientific_name", "family")) %>%
  full_join(prm_hypoxia, by = c("scientific_name", "family"))

predict_gillnet_final <- predict_gillnet_join %>%
  filter(scientific_name %in% prm_hypoxia$scientific_name)

write_csv(predict_gillnet_final, file = here::here("data", "predict_gillnet_final.csv"), na = "")
```

```{r}
# read in the full_threat data
full_threat <- read_csv(here("data", "full_threat_list.csv")) %>%
  filter(!str_detect(scientific_name, "trygon|batos|batus|Dipturus|Raja|Rajella|Gymnura|Dasyatis|Urolophus|Torpedo|Hypanus|Myliobatis|Rhinoptera|Glaucostegus|batis|rhina|Pristis|Anoxypristis|Zapteryx|Hydrolagus|Bathytoshia|Discopyge|Narcine|Styracura|Tetronarce|Urogymnus|Glyphis|Neoharriota|Springeria|Sympterygia|Callorhinchus|Heteronarce|Pastinachus|Taeniura|Neoharriotta|Taeniurops|Okamejei|Narke|Himantura|Rhina|Hongeo|Temera")) %>%
  filter(is.na(avm_pred)) %>%
  select(scientific_name)

# read in the predict data
predict_data_hyp <- read_csv(here("data", "predict_data_hyp.csv")) %>%
  mutate(habitat = case_when(
    str_detect(habitat_associated, "pelagic") ~ "pelagic",
    str_detect(habitat_associated, "demersal") ~ "demersal",
    habitat_associated == "reef-associated" ~ "pelagic"
  )) %>%
  select(estimate, scientific_name, family, habitat, max_size_cm, ventilation_method, reproductive_mode, median_depth, genus, ac, order, source)
```

```{r}
# join datasets
full_threat_join <- full_threat %>%
  full_join(predict_data_hyp, by = "scientific_name") %>%
  left_join(tol) %>%
  select(-genus) %>%
  mutate(species = scientific_name) %>%
  separate_wider_delim(species, names = c("genus", "species"), delim = " ") %>%
  group_by(genus) %>%
  mutate(ac_new_gen = ifelse(is.na(ac), mean(ac, na.rm = TRUE), ac)) %>%
  group_by(family) %>%
  mutate(ac_new_fam = ifelse(is.na(ac), mean(ac, na.rm = TRUE), ac)) %>%
  ungroup() %>%
  mutate(
    ac = ifelse(is.na(ac), ac_new_gen, ac),
    ac = ifelse(is.nan(ac_new_gen), ac_new_fam, ac)
  ) %>%
  mutate(ac = ifelse(scientific_name == "Glaucostegus cemiculus", 5.7727, ac)) %>% # using the value for G. typus from the 'tol' df
  filter(!is.na(ac)) %>%
  select(-ac_new_gen, -ac_new_fam) %>%
  filter(!str_detect(scientific_name, "trygon|Raja|Myliobatis|Glaucostegus|Mobula|raja|Urolophus|Aptychotrema|Akheilos")) # remove remaining oservations from species that shouldn't be predicted

write_csv(full_threat_join, file = here("data", "predict_data_new_iucn.csv"), na = "")
```
