==> src/02_random_forest_longline.Rmd <==
---
title: "random_forest"
author: "Leonardo Feitosa"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
  echo = TRUE,
  message = FALSE,
  warning = FALSE,
  cache = FALSE
)
library(tidyverse)
library(tidymodels)
library(here)
library(randomForest)
library(ranger)
library(googlesheets4)
library(naniar)
library(ggpmisc)
library(patchwork)
library(vip)

set.seed(42)
```

```{r}
prm_elasmo <- read_csv(here("data", "prm_hypoxia.csv")) %>%
  filter(gear_class == "longline") %>%
  filter(!grepl("Rajidae", family)) %>%
  filter(family != "Dasyatidae") %>%
  filter(family != "Pristidae") %>%
  filter(family != "Rhinidae") %>%
  filter(family != "Rhinobatidae") %>%
  filter(family != "Myliobatidae") %>%
  filter(family != "Mobulidae") %>%
  mutate(habitat = case_when(
    habitat_associated == "pelagic" ~ "pelagic",
    habitat_associated == "pelagic-oceanic" ~ "pelagic",
    habitat_associated == "reef-associated" ~ "pelagic",
    habitat_associated == "bathypelagic" ~ "pelagic",
    habitat_associated == "benthopelagic" ~ "demersal",
    habitat_associated == "demersal" ~ "demersal",
    habitat_associated == "bathydemersal" ~ "demersal",
  )) %>%
  mutate(habitat = case_when(
    scientific_name == "Carcharhinus perezi" ~ "pelagic",
    TRUE ~ habitat
  )) %>%
  mutate(ventilation_method = case_when(
    scientific_name == "Ginglymostoma cirratum" ~ "stationary",
    TRUE ~ ventilation_method
  )) %>%
  mutate(reproductive_mode = case_when(
    scientific_name == "Hexanchus griseus" ~ "lecithotrophic viviparity",
    scientific_name == "Hexanchus nakamurai" ~ "lecithotrophic viviparity",
    scientific_name == "Ginglymostoma cirratum" ~ "lecithotrophic oviparity",
    scientific_name == "Orectolobus ornatus" ~ "lecithotrophic viviparity",
    TRUE ~ reproductive_mode
  )) %>%
  mutate(family = case_when(
    scientific_name == "Galeus melastomus" ~ "Scyliorhinidae",
    scientific_name == "Ginglymostoma cirratum" ~ "Ginglymostomatidae",
    scientific_name == "Pseudocarcharias kamoharai" ~ "Pseudocarchariidae",
    TRUE ~ family
  )) %>%
  group_by(scientific_name) %>%
  filter(!grepl("sp", scientific_name)) %>%
  mutate(median_depth = mean(median_depth)) %>%
  mutate(max_size_cm = mean(max_size_cm)) %>%
  distinct()
```

```{r}
# create data subset with variables that will be used in the random forest model
prm_elasmo_subset <- prm_elasmo %>%
  select(
    scientific_name, gear_class, habitat, estimate_type, estimate, sample_size,
    ventilation_method, max_size_cm, median_depth, reproductive_mode, family, ac
  ) %>%
  mutate_if(is.character, as.factor) %>%
  distinct()

# Create subset of the whole dataset with at-vessel mortality and calculate the weighted averages per species and gear
elasmo_avm <- prm_elasmo_subset %>%
  filter(estimate_type == "at-vessel mortality") %>% 
  filter(sample_size > 15) %>% 
  group_by(scientific_name) %>%
  mutate(
    mortality_prop = weighted.mean(estimate, sample_size),
    avm_sd = sd(estimate)
  ) %>% 
  # mutate(mortality_prop = estimate) %>%
  select(ventilation_method, median_depth, max_size_cm, habitat, reproductive_mode, mortality_prop, family, ac) %>% 
  distinct() %>% 
  drop_na()

# Create subset of the whole dataset with post-release mortality and calculate the weighted averages per species and gear
elasmo_prm <- prm_elasmo_subset %>%
  filter(estimate_type == "post-release mortality") %>%
  filter(sample_size > 5) %>%
  group_by(scientific_name) %>%
  mutate(
    mortality_prop = weighted.mean(estimate, sample_size),
    prm_sd = sd(estimate)
  ) %>%
  # mutate(mortality_prop = estimate) %>%
  select(ventilation_method, median_depth, max_size_cm, habitat, reproductive_mode, mortality_prop, family, ac) %>%
  distinct() %>%
  drop_na()
```

```{r}
# split the data into training and test set
## At-vessel mortality
avm_split <- initial_split(elasmo_avm, prop = .7, strata = family)

avm_train <- training(avm_split)
avm_test <- testing(avm_split)

avm_folds <- vfold_cv(avm_train, repeats = 10, v = 5, strata = family)

prm_folds <- vfold_cv(elasmo_prm, repeats = 10, v = 5, strata = family)
```


## Overview

This model aims to predict the likelihood of mortality per taxonomic family between sharks and rays using several species-specific bioecological features such as ventilation mode, median temperature, reproductive mode, median depth, as well as technological features such as the fishing gear each species was captured with.

```{r}
# Create the recipes for each estimate
avm_recipe <- recipe(mortality_prop ~ ventilation_method + median_depth + reproductive_mode + ac, data = avm_train) %>%
  step_normalize(all_numeric_predictors())

prm_recipe <- recipe(mortality_prop ~ median_depth + reproductive_mode + max_size_cm + habitat + ac, data = elasmo_prm) %>%
  step_normalize(all_numeric_predictors())
```

```{r}
# create models
avm_rf <- rand_forest(trees = tune(), mtry = tune(), min_n = tune(), mode = "regression") %>%
  set_engine("ranger", verbose = TRUE, oob.error = TRUE, quantreg = TRUE, importance = "impurity")

prm_rf <- rand_forest(trees = tune(), mtry = tune(), min_n = tune(), mode = "regression") %>%
  set_engine("ranger", verbose = TRUE, oob.error = TRUE, quantreg = TRUE, importance = "impurity")
```

```{r}
# create workflows
avm_wflw <- workflow() %>%
  add_model(avm_rf) %>%
  add_recipe(avm_recipe)

prm_wflw <- workflow() %>%
  add_model(prm_rf) %>%
  add_recipe(prm_recipe)
```

### Cross Validation

Cross validation will be used to choose the best hyperparameters for use in the model: 

* Number of trees
* Number of predictors in each sample
* Minimum data points in each node required for node split

```{r,include=FALSE}
# avm_tune <- avm_wflw %>% # use cross validation to tune parameters
#   tune_grid(
#     resamples = avm_folds, # add cv folds
#     grid = expand.grid(mtry = seq(1, 3, 1), trees = seq(100, 2000, 100), min_n = seq(1, 31, 5))
#   ) # use grid of parameters to test
```

```{r,include=FALSE}
# prm_tune <- prm_wflw %>% # use cross validation to tune parameters
#   tune_grid(
#     resamples = prm_folds, # add cv folds
#     grid = expand.grid(mtry = seq(1, 4, 1), trees = seq(100, 2000, 100), min_n = seq(1, 12, 5))
#   ) # use grid of parameters to test
```

Below displays the results of cross validation tuning

```{r}
# autoplot(avm_tune) + # plot results of tuning
#   theme_bw() # change theme
```

```{r}
# select_best(avm_tune, metric = "rmse", n = 1) # get the best model
```

```{r, include=FALSE}
# avm_final <- finalize_workflow(avm_wflw, select_best(avm_tune, metric = "rmse")) # finalize the workflow with the best model
#
# saveRDS(avm_final, here::here("data", "avm_final.rds"))

avm_final <- readRDS(here::here("data", "avm_final.rds"))
```

```{r, include=FALSE}
avm_fit <- fit(avm_final, avm_train) # fit the data to the training data
```

```{r, include=FALSE}
avm_train_predict <- predict(object = avm_fit, new_data = avm_train) %>% # predict the training set
  bind_cols(avm_train) # bind training set column to prediction

avm_test_predict <- predict(object = avm_fit, new_data = avm_test) %>% # predict the training set
  bind_cols(avm_test) # bind prediction to testing data column
```

The first is the training metrics the second is the testing metrics

```{r}
avm_train_metrics <- avm_train_predict %>%
  metrics(mortality_prop, .pred) # get testing data metrics
avm_train_metrics

avm_test_metrics <- avm_test_predict %>%
  metrics(mortality_prop, .pred) # get testing data metrics
avm_test_metrics
```

```{r}
# pelagic
pelagic_train_metrics <- avm_train_predict %>%
  filter(habitat == "pelagic") %>%
  metrics(mortality_prop, .pred) # get testing data metrics
pelagic_train_metrics

pelagic_test_metrics <- avm_test_predict %>%
  filter(habitat == "pelagic") %>%
  metrics(mortality_prop, .pred) # get testing data metrics
pelagic_test_metrics
```

```{r}
# demersal
demersal_train_metrics <- avm_train_predict %>%
  filter(habitat == "demersal") %>%
  metrics(mortality_prop, .pred) # get testing data metrics
demersal_train_metrics

demersal_test_metrics <- avm_test_predict %>%
  filter(habitat == "demersal") %>%
  metrics(mortality_prop, .pred) # get testing data metrics
demersal_test_metrics
```


Below displays the results of cross validation tuning

```{r}
# autoplot(prm_tune) + # plot results of tuning
#   theme_bw() # change theme
```

```{r}
# select_best(prm_tune, metric = "rmse", n = 1) # get the best model
```

```{r, include=FALSE}
# prm_final <- finalize_workflow(prm_wflw, select_best(prm_tune, metric = "rmse")) # finalize the workflow with the best model
#
# saveRDS(prm_final, here::here("data", "prm_final.rds"))

prm_final <- readRDS(here::here("data", "prm_final.rds"))
```

```{r, include=FALSE}
prm_fit <- fit(prm_final, elasmo_prm) # fit the data to the training data
```

```{r, include=FALSE}
prm_train_predict <- predict(object = prm_fit, new_data = elasmo_prm) %>% # predict the training set
  bind_cols(elasmo_prm) # bind training set column to prediction
```

```{r}
prm_train_metrics <- prm_train_predict %>%
  metrics(mortality_prop, .pred) # get testing data metrics
prm_train_metrics
```

```{r}
my_controls <- control_resamples(
  verbose = TRUE,
  save_pred = TRUE,
  extract = function(x) {
    model <- extract_fit_engine(x)
    model$prediction.error
  }
) # set to get prediction and out of bag errors

set.seed(42)

avm_cv2 <- avm_final %>%
  fit_resamples(avm_folds, control = my_controls) # fit the folds to the final model

set.seed(42)

prm_cv2 <- prm_final %>%
  fit_resamples(prm_folds, control = my_controls) # fit the folds to the final model
```

```{r}
collect_metrics(avm_cv2)

collect_metrics(prm_cv2)
```

## Quantile regression

For AVM

```{r}
preds_bind <- function(data_fit, rf_fit) {
  predict(
    rf_fit$fit$fit$fit,
    workflows::extract_recipe(rf_fit) %>% bake(data_fit),
    type = "quantiles",
    quantiles = c(.05, .25, .5, .75, .95)
  ) %>%
    with(predictions) %>%
    as_tibble() %>%
    set_names(paste0(".pred", c("_05", "_25", "_50", "_75", "_95"))) %>%
    bind_cols(data_fit)
} # function to preform and extract quantile estimates from random forest

avm_data <- full_join(avm_test_predict, avm_train_predict)

# predict quantiles
avm_quant_test <- preds_bind(avm_data, avm_fit) %>%
  mutate(
    range_50 = .pred_75 - .pred_25,
    range_95 = .pred_95 - .pred_05
  )

q1 <- ggplot(avm_quant_test) +
  geom_boxplot(aes(range_50)) +
  theme_bw(base_size = 14) +
  labs(x = "Distribution of IQR for AVM (Longline)") +
  theme(
    axis.title.y = element_blank(),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank()
  )
q1

write_csv(avm_quant_test, here::here("data", "avm_model_predictions.csv"))
```

For PRM

```{r}
# predict quantiles
prm_quant_test <- preds_bind(prm_train_predict, prm_fit) %>%
  mutate(
    range_50 = .pred_75 - .pred_25,
    range_95 = .pred_95 - .pred_05
  )

q2 <- ggplot(prm_quant_test) +
  geom_boxplot(aes(range_50)) +
  theme_bw(base_size = 14) +
  labs(x = "Distribution of IQR for PRM") +
  theme(
    axis.title.y = element_blank(),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank()
  )
q2

write_csv(prm_quant_test, here::here("data", "prm_model_predictions.csv"))
```

## AVM model - real values vs predicted values 

```{r}
p1 <- ggplot(avm_quant_test, aes(x = mortality_prop, y = .pred)) + # plot ln of real versus ln of predicted
  geom_point() +
  stat_poly_line() +
  stat_poly_eq(use_label("eq")) +
  stat_poly_eq(label.y = 0.9) +
  theme_bw(base_size = 14) +
  labs(
    x = "Observed AVM Rate (Longline)",
    y = "Predicted AVM Rate (Longline)"
  )
p1

summary(lm(.pred ~ mortality_prop, data = avm_quant_test))
```

## PRM model - real values vs predicted values

```{r}
p2 <- ggplot(prm_train_predict, aes(x = mortality_prop, y = .pred)) + # plot ln of real versus ln of predicted
  geom_point() +
  stat_poly_line() +
  stat_poly_eq(use_label("eq")) +
  stat_poly_eq(label.y = 0.9) +
  theme_bw(base_size = 14) +
  labs(
    x = "Observed PRM Rate",
    y = "Predicted PRM Rate"
  )
p2

summary(lm(.pred ~ mortality_prop, data = prm_train_predict))
```

### Variable importance for the at-vessel mortality and post-release mortality RF models

```{r}
v1 <- avm_fit %>%
  extract_fit_parsnip() %>%
  vip() +
  theme_bw(base_size = 14) +
  labs(y = "Variable Importance for AVM (Longline)") +
  scale_x_discrete(labels = c("Ventilation Method", "Reproductive Mode", "Hypoxia Tolerance", "Median Depth (m)"))
v1

v2 <- prm_fit %>%
  extract_fit_parsnip() %>%
  vip() +
  theme_bw(base_size = 14) +
  labs(y = "Variable Importance for PRM") +
  scale_x_discrete(labels = c("Habitat", "Reproductive Mode", "Hypoxia Tolerance", "Median Depth", "Maximum Size (cm)"))
v2
```

### Predict additional shark species

```{r}
# filter out rays and missing values
new_dat <- read_csv(here::here("data", "predict_data_new_iucn.csv")) %>%
  # filter(!is.na(median_depth)) %>%
  filter(!is.na(ac)) %>%
  filter(!is.na(ventilation_method)) %>%
  filter(!is.na(reproductive_mode)) %>%
  filter(!is.na(habitat)) %>%
  filter(!is.na(max_size_cm)) %>%
  group_by(genus) %>%
  mutate(median_depth_new = mean(median_depth, na.rm = TRUE)) %>%
  ungroup() %>%
  mutate(median_depth = ifelse(is.na(median_depth), median_depth_new, median_depth)) %>%
  mutate(ventilation_method = as.factor(ventilation_method)) %>%
  mutate(habitat = as.factor(habitat)) %>%
  mutate(reproductive_mode = as.factor(reproductive_mode)) %>%
  mutate(reproductive_mode = as.factor(reproductive_mode))

# predict AVM
predict_data <- list(avm_train, avm_test, new_dat) %>%
  reduce(full_join) %>%
  rename(avm_mort = mortality_prop) %>%
  full_join(elasmo_prm) %>%
  rename(prm_mort = mortality_prop) %>%
  distinct() %>%
  mutate(habitat = as.factor(habitat)) %>%
  filter(!is.na(ac))

predict_avm <- predict(object = avm_fit, new_data = predict_data) %>% # predict the training set
  bind_cols(predict_data) %>% # bind training set column to prediction
  rename(avm_pred = .pred)

# predict PRM
predict_avm_prm <- predict(object = prm_fit, new_data = predict_avm) %>% # predict the training set
  bind_cols(predict_avm) %>% # bind training set column to prediction
  rename(prm_pred = .pred)

# combine predictions
full_prediction_avm <- preds_bind(predict_avm_prm, avm_fit) %>%
  rename(
    avm_75 = .pred_75,
    avm_95 = .pred_95,
    avm_05 = .pred_05,
    avm_25 = .pred_25,
    avm_50 = .pred_50
  )

full_predictions <- preds_bind(full_prediction_avm, prm_fit) %>%
  rename(
    prm_75 = .pred_75,
    prm_95 = .pred_95,
    prm_05 = .pred_05,
    prm_25 = .pred_25,
    prm_50 = .pred_50
  ) %>%
  select(
    order, genus, prm_05, prm_25, prm_50, prm_75,
    prm_95, avm_05, avm_25, avm_50,
    avm_75, avm_95, prm_pred, avm_pred,
    scientific_name, ventilation_method, median_depth,
    max_size_cm, habitat, reproductive_mode,
    ac, avm_mort, family, prm_mort
  )

write_csv(full_predictions, here::here("data", "longline_model_predictions.csv"))
```
